+++ frontend/src/pages/Admin.tsx
@@ -5,7 +5,6 @@ export const Admin = () => {
   return (
     <>
       <>
-        <h1>Admin</h1>
         <Outlet />
       </>
     </>