+++ frontend/src/pages/Admin/index.tsx
@@ -5,7 +5,6 @@ export const Admin = () => {
   return (
     <>
       <>
-        <h2>Admin</h2>
         <Outlet />
       </>
     </>